ai_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-generativeai  # Dùng cho Gemini như trong requirement
chromadb             # Cho RAG
python-multipart